modules/derma.py
import streamlit as st
import pandas as pd
import time
from PIL import Image
import numpy as np

def show_derma():
    st.title("🤳 Derma-Scan: Vision-Based Wound & Mole Tracker")
    st.markdown("""
    This module uses **Computer Vision** to monitor skin health. 
    It tracks wound healing progress and analyzes suspicious moles using the **ABCDE** clinical criteria.
    """)

    tab1, tab2 = st.tabs(["🩹 Wound Tracking (Diabetic Foot)", "🔍 Mole Analysis (Skin Cancer)"])

    with tab1:
        st.subheader("Diabetic Wound Progression")
        st.write("Upload a photo of the wound to analyze healing rate and infection signs.")
        
        img_file = st.file_uploader("Upload Wound Photo", type=["jpg", "png", "jpeg"], key="wound_up")
        
        if img_file:
            st.image(img_file, caption="Current Wound State", use_container_width=True)
            if st.button("Analyze Wound Healing"):
                with st.spinner("Calculating tissue area and vascularization..."):
                    time.sleep(2)
                
                # İyileşme Verileri
                st.success("Analysis Complete!")
                c1, c2 = st.columns(2)
                c1.metric("Wound Area", "4.2 cm²", "-0.5 cm² (Improving)")
                c2.metric("Infection Risk", "Low", "Normal Temp")
                
                # İyileşme Grafiği
                st.write("**Healing Progress Over 4 Weeks**")
                healing_data = pd.DataFrame({
                    'Week': [1, 2, 3, 4],
                    'Area (cm²)': [5.8, 5.2, 4.7, 4.2]
                })
                st.line_chart(healing_data.set_index('Week'))
                st.info("💡 **Clinical Note:** Granulation tissue is healthy. Keep current dressing protocol.")

    with tab2:
        st.subheader("ABCDE Mole Assessment")
        st.write("Upload a photo of a mole to check for clinical warning signs of Melanoma.")
        
        mole_file = st.file_uploader("Upload Mole Photo", type=["jpg", "png", "jpeg"], key="mole_up")
        
        if mole_file:
            st.image(mole_file, caption="Analyzing Surface Patterns...", use_container_width=True)
            
            st.write("### AI Risk Scoring")
            # Simüle edilmiş ABCDE Analizi
            st.table(pd.DataFrame({
                'Criterion': ['Asymmetry', 'Border', 'Color', 'Diameter', 'Evolution'],
                'Result': ['Regular', 'Sharp', 'Uniform', '4mm', 'Stable'],
                'Risk': ['Low', 'Low', 'Low', 'Low', 'Low']
            }))
            
            st.success("✅ Overall Assessment: Low Risk. Suggest annual check-up.")