requirement.txt
streamlit>=1.25.0
pandas>=2.0.0
matplotlib>=3.7.0
numpy>=1.24.0
Pillow>=9.5.0
reportlab>=4.0.0
# OpenCV is used for Future Computer Vision Modules
opencv-python-headless>=4.8.0